Adapts type check for boolean columns in ClickHouse. ClickHouse 0.2.6 internally stores Booleans as TinyInts instead of using full-integer columns
Fixes: SIRI-539

+++ src/main/java/sirius/db/jdbc/schema/ClickhouseDatabaseDialect.java
@@ -173,7 +173,7 @@ public class ClickhouseDatabaseDialect extends BasicDatabaseDialect {
         if (type == other) {
             return true;
         }
-        if (in(type, other, Types.BOOLEAN, Types.INTEGER)) {
+        if (in(type, other, Types.BOOLEAN, Types.TINYINT)) {
             return true;
         }
         if (in(type, other, Types.VARCHAR, Types.CHAR)) {
